- added check_events function - added do_events function and added support for LSA max age event - corrected ospf0->cost to uint16_t instead of uint8_t - fixed interface remove function to handle event list as nlist instead of list

+++ ospf-replay/replay.h
@@ -70,7 +70,7 @@ struct ospf {
 	int hello_interval, dead_interval, transmit_delay, retransmit_interval;
 
 	// global cost setting
-	u_int8_t cost;
+	u_int16_t cost;
 
 	// passive interface default
 	u_int8_t passif;
